introduce query.Flat
This introduces a new query.Flat type that makes currently implicit
constraints more explicit and updates NewFlatJob to use that type.
Note that this doesn't update many of the calls in ToFlatJob to use
query.Basic, and just relies on the fact that a query.Flat can be
infallibly converted to a query.Basic with ToBasic().