repo-updater: Add the ability to manually purge repos (#35107)
Add very basic ability to manually purge the oldest deleted or blocked repos
from an instance. This is initially only intended for cloud as we have a purge
worker that runs on customer instances.

For safety you have to specify a limit and we default to purging 1 repo per
second, but this is configurable via a query parameter.

Once we've cleaned up repos on cloud, we can consider enabling the worker on
cloud too.

After scheduling some repos to be deleted we can see progress using the
src_repoupdater_purge_success and src_repoupdater_purge_failed counters.