github: shorten test plan description in template (#35300)
I want to make it easier to delete the multiple lines of test plan
reminder. Now that devs are aware of this change, we can use a more
succinct message. The link to testing principles explains what a test
plan is, which can help new developers.

Test Plan: n/a