dev/ci: make sg lint in CI conditional, unify linter steps (#35331)
Introduces changed.GetLintTargets which is used to decide what sg lint gets run. We assert all lints are valid with a unit test that pulls from dev/sg/linters. Ideally the condition would be inverted, with linters defining changed.Diff types to proc on, but right now changed is an enterprise package which we can't import from the OSS sg.