Reopen "optimize by construction" (#35487)
In order to create queries that are optimized for Zoekt, we currently run an optimization pass over the generated j tree, regenerating jobs that can use a query.Basic directly and replacing them with a NoopJob.

This modifies that flow to instead generate jobs before splitting a query.Basic into query.Flat. The basic idea is that now NewBasicJob generates jobs directly if those jobs can use the structure of a basic query. Otherwise, we split the basic query into flat queries (formalized as query.Flat in the followup PR) and create jobs from flat queries then add them to the basic job.