remove references to line preview from client code
Currently, for each line match, we send the content of the line the
match belongs to. I'm planning to get rid of that since we'd like
matches to be able to span multiple lines, so having a preview of the
line that is matched no longer makes sense.

In the web client, getting rid of it is easy because it is never used,
so this just deletes any references to it without removing it from the
API.