Add group.OrderedParallel (#35572)
This mints a new lib/group package and creates group.OrderedParallel.

The purpose of group.OrderedParallel is to solve the problem of executing an expensive task
on every event in an ordered stream. This has come up a few times in search backend code, but
generics now allow us to solve this in a reuseable and more performant mannner.