gitserver: Properly set clone state after removing a repo (#35588)
Previously we always received the undeleted repo name which meant that
when we tried to update the clone state by repo name it would not find a
match since the name in the database inludes the 'DELETED-' prefix.

Now, we only pass the undeleted name to addForRepo in order to map to the
correct gitserver instance. From that point we pass around the full
"deleted" name so that we can update the status correctly.