dependencies: Do not hard fail when a commit cannot be resolved
A typical case is an organization that has empty repositories (say,
useful because of another code host feature, such as an issue tracker).

We would resolve the commits from valid repos but fail
to resolve the commits from empty repos, and since our usage of the
resolver did not ignore the errors, we would fail.

Whether the ignoreErrors flag in GetCommits should exist at all,
now that we have removed its only client, is a matter of a separate PR.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>