repos: Make a best effort to update external service timestamps (#35697)
There were certain error cases before where we would just return an
error and not update the timestamp on the external service. These would
lead to the last_synced_at timestamp growing older and older until it
triggered an alert. It would also cause us not to backoff when syncing
these problematic services.