codemirror: Fix crash caused by token highlight (#35704)
In hindsight this bug is "obvious" but it was quite obscure first. The core problem is that the extension which is responsible for highlighting the currently hovered token holds onto the token object even if token doesn't exist anymore. Should the token refer to a position outside of the new/current document, an error is thrown. This only happens when the user hovers over a token and then continues to edit the input without moving mouse: We only cleared the token when the mouse leaves the input.

There are two parts to the solution:

1. Instead of storing the hovered token, I'm just storing the hover position and look up the token when the decoration is created. This prevents storing stale data in the extension. Even if the stored position is invalid, looking up a token at that position wouldn't yield any result and thus wouldn't create a decoration.
2. The hover position is now cleared when the document changes, i.e. the token highlight is removed when the input changes. This replicates Monaco's behavior. I also made the necessary changes to also hide the tooltip information.