Search backend: Unify job tracing in observe.go (#35685)
* Add Tags() method to jobs without traceable fields
Some jobs have only a single field which is their child job tree. We want these jobs to continue to implement observableJob so that we can pass them to job.StartSpan(). This commit adds a Tags() method to these jobs which returns an empty slice, in preparation for adding Tags() to the observableJob interface.

* Add Tags() to observableJob interface and add fields to spans in job.StartSpan()
All jobs now have their fields added to their spans in job.StartSpan().

* Merge observableJob into Job
Add Tags() method on the Job interface instead of having observableJob as a separate interface. This more clearly indicates that all jobs should be observable.