migration: add a trigger to repo table to insert gitserver_repo (#35633)
This change removes the possibility of having an inconsistent number of rows in repo and gitserver_repos tables. As soon as new repo is added, the trigger inserts a gitserver repo to the corresponding table making the number of repos and gitserver_repos strictly consistent. It used to be eventually consistent because it took some time for gitserver to clone the repo and add an entry to gitserver_repos table after the repo was added to repo table.