search: rename searcher.SearcherJob to searcher.Job
This is to remove the stutter. This commit is more of an RFC. If people
like the removal of the stutter, I will go ahead and rename most of the
jobs we have to avoid the stutter.

The description is all that really needs reviewing. The code is just the
output of gopls renaming the symbol.

Test Plan: go test ./internal/search/...