Search backend: update searcher client to accept multiline matches (#35897)
This updates the types in the searcher protocol to have an optional
multiline return type. Right now, the client will accept both
single-line or multiline results, re-interpreting single-line as the
multiline type. Once we merge this and roll searcher out, we can update
the server to actually send multiline matches.