fix search result match range highlighting after a unicode character Change `highlightNode.ts` to highlight ranges based on code point index instead of byte index.
In `highlightNode.ts`, string indexing is used to highlight match ranges in search results. However, the range indexes coming from the search backend are counted in terms of runes, whereas string indexing is based on bytes.
JavaScript uses UTF-16 string encoding, which encodes Unicode code points as either one or two hex characters. This was causing highlight ranges to be offset by one when the range came after a two-hex-character-encoded Unicode code point, such as certain emojis. Unpacking a string into an array of code points before slicing the range to be highlighted fixes this issue.