Use `pip/` instead of `python/` for Python package repos
Following up from
https://sourcegraph.slack.com/archives/C032G1CLM7X/p1652957835148599.
The `python/` prefix is undesirable because makes it confusing
to know the difference between `repo:^python` and `lang:python`.
With `python/` as the prefix it's natural to use the phrase "python repository"
to explain such repos while the prefix is only used for "pip installable
package repos".

For consistency with the `npm/` and `maven/` prefixes, the `pip/` prefix
works better than `python/`.