searcher: add ObservationContext to diskcache
ObservationContext on diskcache was introduced in 5ac235a5f. This means
for about 5 months we haven't had metrics for it on searcher. This
commit is the most minimal introduction for it. We don't yet integrate
ObservationContext into searcher.

I considered just constructing an ObservationContext in Start, but then
the prometheus registry would panic in tests due to multiple
constructions of it.

Test Plan: go test and running an unindexed search locally.