diskcache: include all files in size calculation
Kubernetes evicts searchers if they go too far over the cache size. This
means we need to maintain our cache size for all files, not just zip
files. We can only safely remove zip files, so we still only remove
those. But when making eviction decisions, it is now based on the total
size of the cache dir, not just the size of zip files within.

Test Plan: added a unit test. It was useful since info.Size() works on a
directory but we want to ignore that.