SearchBackend: refine HunkMatch to include PreviewStart (#36176)
This refines the new HunkMatch type to include PreviewStart so that
we have both the line number and the byte offset of Preview available
for further processing. PreviewEnd can always be calculated from
Preview, so I felt it was best to just provide a start location rather
than a range that is represented by Preview.