Search backend: add comments to HunkMatch and MatchedContent (#36179)
This adds a bunch of comments to HunkMatch since I think I'm happy with
where this design landed. Additionally, it adds MatchedContent, which
returns the exact set of strings matched by the query. It's not used
except in tests yet, but I wanted to add that now to pin that behavior
since that's critical to the design of what we store in HunkMatch.