searcher: prefix all prometheus variables with metric (#36161)
This is a pattern we started using a while ago, but the code in searcher
predates that. This was done via gopls doing renames.

Test Plan: go test

Accidently squashed merged into this commit, for convenience just combinging:

* searcher: prometheus gauge for SEARCHER_CACHE_SIZE_MB (#36162)

This is to allow us to create a dashboard which includes the configured
maximum size of the cache or make a percentage of cache used. This will
make it easier to check admin configurations, since we don't need to
check the configuration we can just check the dashboard.

Test Plan: started locally and checked the value of
searcher_store_max_cache_size_bytes in the metrics page for searcher.