gitserver: remove ArchiveURL (#36243)
The only user of this was an internal endpoint which used to be used by
zoekt. Zoekt removed its use of Sourcegraph archive a few months ago so
it is no longer needed. Additionally, I confirmed on sourcegraph.com
that the route had not been used in the last 30 days via the prometheus
query

  max_over_time(src_http_requests_last_timestamp_unixtime{route="internal.git.tar"}[30d])

Test Plan: CI