JetBrains: Improve key handlers and handle alt+enter in the preview panel (#36295)
This PR fixes key handlers that currently sometimes get swallowed when the editor has focus. It seems that in this case, the manual swing listener _never fires_ which leads me to believe that the editor has a custom key handling (necessary for the advanced text editing capabilities). To work around this, we now add a global key listener. However since these are global, we need to be careful to only respond to events if they happen in the right project. To do that, we check if the popover is active and shown. The test plan includes tests with two projects open side-by-side to ensure there are no issues in this setup.

Additionally, since I've restructured the key handler code in Java, it was easy for me to add support for `alt+enter` when the editor is in focus as well.