repo-updater: Return refreshed token
Before, we would update it in the database but not actually return the
new token which meant that the client we create would still use the OLD
token. This change updates the token in the database and also returns
the new token so that it can be used right away.