httpapi: handle empty IndexedIDs
This fixes a bug where frontend uses the list of IDs of indexed repos provided
by Zoekt as a filter for a call to the repo store.

However if Zoekt provides an empty list, the filter is empty and the
repo store returns all repos, instead of returning none.

This led to frontent returning a list of repos that should be indexed by
Zoekt + a list of all repos that have not yet been indexed by any other
Zoekt.