dev/ci: run docker-images builds if docker-images has been changed (#36425)
If a diff in docker-images is detected, we run a build of all images. We just build everything for simplicity here - if we change this build to happen via a sg command, we can make it diff-based, but these days we intentionally do not make individual changed files available during pipeline generation, so I think this is a reasonable approach.