Search backend: simplify newToMatches (#36433)
This is just a small style change. At some point I'd implemented this as
a function that returns a closure over its arguments, but I think that
was before we had streaming for searcher so we weren't already creating
a closure here. In any case, when pairing with Thomas, he rightly asked
why we did it this weird and convoluted way and I didn't have a real
answer, so I'm making it less confusing.