monitoring(repo-updater): add duration on critical rate limit alerts, update next steps (#36683)
The existing advice is not great, because restarting the pod doesn't guarantee a new public IP, and it's possible customers are using a public gateway (e.g. Cloud NAT) with VM without a public IP address. It's also noted that this alert comes up quite frequently - I think on larger instances running up against the limit is quite common and even expected, and Sourcegraph should for the most part continue working even if it exhausts its rate limits. It becomes a critical issue only if the rate limit is exhausted immediately after a rate limit reset. We can detect this by checking to see if the rate limit is below the threshold for ~most of a window.

This change addresses both issues with updates to the alerts and NextSteps.