executor: Implement shallow clones and sparse checkouts
To improve performance, we now offer these two clone modes. They will be used in a follow up PR by SSBC to replace repo zip archive fetching with proper clones, that are faster because git can stream, unlike zip.

This also changes how local workspaces are set up: the repo is now cloned into a subdirectory. That helps with a problem where `VirtualMachineFiles` would live in the repo, cluttering it.
For SSBC, I just had to fix up one path (for the better)
For code intel, it appears to me that you aren't affected by this change? The WD is set properly to the new path, so it should be an opaque change and you don't use `VirtualMachineFiles`.
Ideally, we do a follow-up PR that resets the WD to `/` by default but allows to specify a custom wd? Then code intel could always hard-code it to `repo`. Anyways, this feels like a good improvement. It is much faster for me locally.
I think I heard you say that you need full clones for LSIF so code intel doesn't benefit too much from this yet, but maybe there's potential to use sparse checkouts instead? Or we can find a way to get it to work without the full repo, this brings down IO for sg/sg from ~800MB to ~52MB.