fix concurrent transactions
This fixes an issue I was seeing with creating code monitors for
multiple repos where we'd get weird `conn busy` errors. Basically, the
issue is that transactions can't be used concurrently and the
`database.DB` passed into this function is a transaction in
`CreateCodeMonitor`. The jobs run in parallel, which triggers these
errors sporadically.