search: Fix React errors about unsupported DOM props
[#36204](https://github.com/sourcegraph/sourcegraph/pull/36204/files)
added a change to add all "rest" props to a <div> element, which had the
effect that a lot of component specific props have been added to the DOM
element, and React complains about that.

This commit fixes this in two ways: Only the props required by
`<Toggles>` are passed and the rest props spread is removed (it's not
clear to me why this was needed in the first place).