backend: mint `InsightsDB` interface (#36890)
This commit mints an edb.InsightsDB type that plays a similar role to the database.DB type. This also replaces many instances where we use dbutil.DB for the code insights database with this named type. It's basically just a lightweight, typed wrapper around dbutil.DB that is also a ShareableStore.