JetBrains: Force hide the material UI overlay when the Sourcegraph window is hidden (#36905)
Another attempt at fixing #36479. This time it does not bring back the web view issues on macOS.

We now manually emit an action defined by the material UI theme to hide the overlay it opens whenever a popover is
created. This third-party plugin does not work with our approach of keeping the popover alive and thus, when the
Sourcegraph popover is closed, their custom overlay stays active.

Background on how we came up with the solution can be found here: mallowigi/material-theme-issues#179 (comment)

Note: This is not an ideal fix yet as the overlay does not re-appear when the window is re-opened. Given that this is not the only issue with the overlay and the fact that I was able to find more by using this theme myself for a bit, I am not super worried about this.