frontend: Use generic timeout in integration test (#36900)
While working on integration tests for the search query input and trying
to figure out how to handle waiting on the editor if we want to be able
to test different editor implementations (Monaco/CodeMirror), I noticed
that the timeout for waiting on the editor gives other components time
to "settle".
This became evident after I removed one of the editor checks.

But that means that irrespectively of whether or not an editor is
rendered on the page, there are pages with components which need time to
settle. Using a generic timeout (instead of waiting for something that
doesn't exist anyway) expresses that intent better. And that generic
timeout also will leave enough time for the editor(s) to update.

I also flipped the order in which we take screenshots so that we don't
have to switch back the color scheme (and wait for that) again (the
assumption seems to be that light mode is default).