Code monitors: enable repo-aware monitors by default (#37028)
This enables the cc-repo-aware-monitors flag by default if it is not
set. This will make it so when customers update, they will be using
repo-aware monitors. I intend to leave the feature flag in place until
after the release cut so we have an emergency escape hatch, but I don't
expect we'll need it.