Structural search: Allow callers to access comby's stdin and stdout pipes (#37083)
* wip

* Add ability to access comby pipes
This commit splits the `PipeTo()` method into `SetupCmdWithPipes()` and `StartAndWaitForCompletion()`. `SetupCmdWithPipes()` will return an `*exec.Cmd` pointing to `combyPath` with the provided args, as well as pipes connected to the command's stdin, stdout, and stderr. `StartAndWaitForCompletion()` starts the command and waits for it to run to completion.
This change allows callers of the `comby` package the ability to access pipes to comby's stdin and stout, which enables both writing to stdin and reading from stdout in a streaming manner. This ability is needed for implementing streaming structural search in a future PR.

* revert introduction of tar as a comby input type
Introducing tar files as a valid input into comby will be more relevant to a future PR where their use will actually be implemented, rather than in this one

* slightly more accurate log statements

* create command with context

* simplify comby.go

* rewrite pipeTo() as a utility function for tests

* rename pipeTo() to runWithoutPipes()

* move runWithoutPipes() to comby_test.go