searcher: do not log net.OpError
I was watching searcher logs and noticed huge amounts of log spam when a
connection is closed. This is since we can try to find several matches
before we check if context is canceled. We then get a log message per
match informing us of something like:

  write tcp 10.164.18.30:3181->10.164.27.49:56316: write: broken pipe

Test Plan: To reproduce I made hybrid search in zoekt write ^10 more
matches, then added some timing which called ts.Config.Close(). This
reproduced the error. Additionally I think the error checking code is a
little tricky, so I introduced a test specifically for the helper.