Event logs: allow logging URLs from all sourcegraph domains in local event logs (#37236)
This updates the local event log URL redaction code to allow any URLs
ending in sourcegraph.com to be logged. Note that when sending event
logs to BigQuery, we redact all https://sourcegraph.com URLs so that
private information isn't leaving the cluster, however that isn't a
concern for local event logs.