ssbc: Don't keep execution results, recreate from steps (#36758)
This drops execution cache results from SSBC entirely (not step results, just execution) and only uses step cache results instead. The cache LRU swaps them out eventually so we don't need a costly migration. This is accomplished by porting the logic to convert a step cache result into changeset specs in the backend. No executor run is required for the scenario where we have all step entries but no execution result anymore due to cache eviction. We reduce data duplication in the database.
Most notably, check that I adjusted the test not to include a cache entry log line anymore, but all the changeset specs are still generated, just as before.

As a side-effect, this closes #36540.