gitlab: Refresh expired accounts (#37260)
We ensure that even if a GitLab token expires, we still try to refresh it.

It was possible in the previous code that we would attempt to refresh the token
with a small 10 second grace period and then it would take us more than 10
seconds to actually use the token. This could cause the token to be expired by
the time we use it which would then cause us to mark the account as expired and
never try to refresh it again.